services/ocr_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
easyocr==1.7.1
requests==2.31.0
